Update QueryBuilder get method to run getFresh method
Following the latest update of the laravel framework (5.1.20), the BaseQuery builder that used to just run the getFresh method (overriden by the mongodb child builder) in the get function, now calls the runSelect method that relies on the mysql Grammar object being present that this builder does not need to run.

This update simply removes the parent get() call and directly calls it's own getFresh method, thus return the functionality to what it was before the laravel update.